Simplify RouteConfigurationTest
The first version of this test uses various Spring annontation which
starts up a Spring environment in order to run the test. This can be
useful in some circumstances, but is not needed here.

This updated implementation only dependency is WebTestClient, which
provides synthetic request and response objects. The test is more
explicit and should run faster as it does not need a Spring context.