Add RouteConfigurationTest
Tests that a request to the /greet route returns a 200 with the expected
greeting.

Note the need to convert a nullable byte stream before converting to a
string in the test (!! operator). This is one of the rougher interfaces
between the null safe world of Kotlin and the world of Spring and Java.

An unsafe type conversion is not ideal, and should not be included in
production code, but feels a reasonable compromise in a test scenario.
As the test would not pass if this code through an error, and writing
this in a type safe way would introduce cognitive overhead that would
take focus away from what is actually being tested.